% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{add_project_assignments}
\alias{add_project_assignments}
\title{Add staff assignments to schedulke}
\usage{
add_project_assignments(schedule, workplan)
}
\arguments{
\item{db_name}{The name of the database to create}

\item{tmp}{schedule}
}
\description{
Add staff assignments to schedulke
}
\keyword{internal}
