% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get-schedule.R
\name{add_project_calendar}
\alias{add_project_calendar}
\title{create a list of employees that are to be assigned to projects}
\usage{
add_project_calendar(tmp, db_name)
}
\arguments{
\item{tmp}{schedule}

\item{db_name}{The name of the database to create}
}
\description{
create a list of employees that are to be assigned to projects
}
\keyword{internal}
