% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get-schedule.R
\name{add_staff_assignment}
\alias{add_staff_assignment}
\title{Add staff assignments to schedulke}
\usage{
add_staff_assignment(tmp, db_name)
}
\arguments{
\item{tmp}{schedule}

\item{db_name}{The name of the database to create}
}
\description{
Add staff assignments to schedulke
}
\keyword{internal}
