% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get-schedule.R
\name{add_staff_out_of_office}
\alias{add_staff_out_of_office}
\title{Add out of office times to schedule}
\usage{
add_staff_out_of_office(tmp, db_name)
}
\arguments{
\item{tmp}{schedule}

\item{db_name}{The name of the database to create}
}
\description{
Add out of office times to schedule
}
\keyword{internal}
