% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,leave-method}
\alias{as.data.frame,leave-method}
\title{Coerce Object leave to a data frame}
\usage{
\S4method{as.data.frame}{leave}(x)
}
\arguments{
\item{x}{A \code{leave} object.}
}
\description{
Coerce Object leave to a data frame, avoiding using the "slot" notation.
}
