% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{as.data.frame,project_phases-method}
\alias{as.data.frame,project_phases-method}
\title{Coerce Object phases to a data frame}
\usage{
\S4method{as.data.frame}{project_phases}(x)
}
\arguments{
\item{x}{A \code{phases} object.}
}
\description{
Coerce Object phases to a data frame, avoiding using the "slot" notation.
}
