% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{as.data.frame,project_roles-method}
\alias{as.data.frame,project_roles-method}
\title{Coerce Object roles to a data frame}
\usage{
\S4method{as.data.frame}{project_roles}(x)
}
\arguments{
\item{x}{A \code{role} object.}
}
\description{
Coerce Object roles to a data frame, avoiding using the "slot" notation.
}
