% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,projects-method}
\alias{as.data.frame,projects-method}
\title{Coerce Object project to a data frame}
\usage{
\S4method{as.data.frame}{projects}(x)
}
\arguments{
\item{x}{A \code{projects} object.}
}
\description{
Coerce Object project to a data frame, avoiding using the "slot" notation.
}
