% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{as.data.frame,public_holidays-method}
\alias{as.data.frame,public_holidays-method}
\title{Coerce Object public_holidays to a data frame}
\usage{
\S4method{as.data.frame}{public_holidays}(x)
}
\arguments{
\item{x}{A \code{public_holidays} object.}
}
\description{
Coerce Object public_holidays to a data frame, avoiding using the "slot" notation.
}
