% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,resources-method}
\alias{as.data.frame,resources-method}
\title{Coerce Object resource to a data frame}
\usage{
\S4method{as.data.frame}{resources}(x)
}
\arguments{
\item{x}{A \code{resource} object.}
}
\description{
Coerce Object resource to a data frame, avoiding using the "slot" notation.
}
