% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,responsibilities-method}
\alias{as.data.frame,responsibilities-method}
\title{Coerce Object responsibilities to a data frame}
\usage{
\S4method{as.data.frame}{responsibilities}(x)
}
\arguments{
\item{x}{A \code{responsibilities} object.}
}
\description{
Coerce Object responsibilities to a data frame, avoiding using the "slot" notation.
}
