% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,roles-method}
\alias{as.data.frame,roles-method}
\title{Coerce Object roles to a data frame}
\usage{
\S4method{as.data.frame}{roles}(x)
}
\arguments{
\item{x}{A \code{phases} object.}
}
\description{
Coerce Object roles to a data frame, avoiding using the "slot" notation.
}
