% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,staff_schedule-method}
\alias{as.data.frame,staff_schedule-method}
\title{Coerce Object staff_schedule to a data frame}
\usage{
\S4method{as.data.frame}{staff_schedule}(x)
}
\arguments{
\item{x}{A \code{staff_schedule} object.}
}
\description{
Coerce Object staff_schedule to a data frame, avoiding using the "slot" notation.
}
