% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.data.frame,time_estimates-method}
\alias{as.data.frame,time_estimates-method}
\title{Coerce Object time_estimates to a data frame}
\usage{
\S4method{as.data.frame}{time_estimates}(x)
}
\arguments{
\item{x}{A \code{time_estimates} object.}
}
\description{
Coerce Object time_estimates to a data frame, avoiding using the "slot" notation.
}
\keyword{internal}
