% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{as.list,workplan-method}
\alias{as.list,workplan-method}
\title{Coerce Object workplan to a list of data frames}
\usage{
\S4method{as.list}{workplan}(x)
}
\arguments{
\item{x}{A \code{workplan} object.}
}
\description{
Coerce Object workplan to a list of data frames, avoiding using the "slot" notation.
}
\keyword{internal}
