% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-assign_and_remove.R
\name{assign_staff}
\alias{assign_staff}
\title{Assign staff member to project phase}
\usage{
assign_staff(workplan, staff_name, project_name, project_role_name,
  staff_contribution)
}
\arguments{
\item{workplan}{The name of the database to create}

\item{staff_name}{staff name}

\item{project_name}{project name}

\item{project_role_name}{project_phase_name}

\item{staff_contribution}{staff con}
}
\description{
Assign staff member to project phase
}
