% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-workplan_io.R
\name{build_sample_workplan}
\alias{build_sample_workplan}
\title{Create a random workplan}
\usage{
build_sample_workplan(excel_file_name = "my-workplan.xlsx")
}
\arguments{
\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
\examples{
library(workplanr)
}
