% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{calculate_start_and_end_dates}
\alias{calculate_start_and_end_dates}
\title{Initialise schedule table}
\usage{
calculate_start_and_end_dates(workplan)
}
\arguments{
\item{workplan}{The name of the database to create}
}
\description{
Initialise schedule table
}
\keyword{internal}
