% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{calculate_workplan}
\alias{calculate_workplan}
\title{Add staff assignments to schedulke}
\usage{
calculate_workplan(workplan)
}
\arguments{
\item{workplan}{schedule}
}
\description{
Add staff assignments to schedulke
}
\keyword{internal}
