% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_workplan.R
\name{correct_classes}
\alias{correct_classes}
\title{Create Excel file for project inputs}
\usage{
correct_classes(staff, staff_capacity, projects, project_probability,
  project_start, project_end, project_phases, project_time_estimates,
  staff_on_leave, leave_start, leave_end, leave_description,
  public_holidays_date, public_holidays_name,
  staff_project_assignment_capacity)
}
\arguments{
\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
\examples{
library(workplanr)
}
\keyword{internal}
