% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-read-base-project-inputs.R
\name{create.project.template}
\alias{create.project.template}
\title{ggplot colours}
\usage{
create.project.template(file = "my-project-template-inputs.xlsx")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
