% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-create-random-project.R
\name{create.random.project}
\alias{create.random.project}
\title{Creates a randomised project management problem}
\usage{
create.random.project()
}
\arguments{
\item{nothing}{}
}
\description{
Creates a randomised project management problem
}
