% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcast-schedule.R
\name{create.team.plan}
\alias{create.team.plan}
\title{Creates a randomised project management problem}
\usage{
create.team.plan(my.work.plan, resources)
}
\arguments{
\item{nothing}{}
}
\description{
Creates a randomised project management problem
}
