% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcast-schedule.R
\name{create.work.plan}
\alias{create.work.plan}
\title{Creates a randomised project management problem}
\usage{
create.work.plan(resources, projects, time.estimates, public.holidays,
  project.teams, responsibilities, leave)
}
\arguments{
\item{nothing}{}
}
\description{
Creates a randomised project management problem
}
