% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-create_new_workplan.R
\name{create_new_workplan}
\alias{create_new_workplan}
\title{Populate workplanr database with data}
\usage{
create_new_workplan(staff, projects, project_phases, project_roles,
  out_of_office, public_holidays, time_estimates, roles_responsibilities,
  staff_name_for_unassigned_work = "unassigned")
}
\arguments{
\item{staff}{Names of staff members}

\item{projects}{Names of projects}

\item{project_phases}{List of phases in any project in order of execution}

\item{project_roles}{List of project roles}

\item{out_of_office}{Names of staff that are going to be out of the office}

\item{public_holidays}{A data frame of dates of public holidays}

\item{time_estimates}{Time estimates of how long each phase will take in relation}

\item{roles_responsibilities}{Data frame for responsibilites of roles across project phases}

\item{staff_name_for_unassigned_work}{Dummy staff member to assign work to}
}
\description{
Populate workplanr database with data
}
