% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-create_new_workplan_db.R
\name{create_new_workplan_db}
\alias{create_new_workplan_db}
\title{create a list of employees that are to be assigned to projects}
\usage{
create_new_workplan_db(staff, projects, calendar, project_phases,
  out_of_office, public_holidays, time_estimates, project_assignments,
  db_name = "my_workplan.sqlite")
}
\arguments{
\item{staff}{Names of staff members}

\item{projects}{Names of projects}

\item{calendar}{Dates of work plan}

\item{project_phases}{List of phases in any project in order of execution}

\item{out_of_office}{Names of staff that are going to be out of the office}

\item{public_holidays}{A data frame of dates of public holidays}

\item{time_estimates}{Time estimates of how long each phase will take in relation}

\item{project_assignmentsAmount}{of time each staff is expected to dedicate to each [project, phase]}
}
\value{
TRUE if script exectues completely
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
library(workplanr)
}
