% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-phases.R
\name{create_phases}
\alias{create_phases}
\title{create a list of phases in any project}
\usage{
create_phases(phases)
}
\arguments{
\item{phases}{List of phases in any project in order of execution}
}
\value{
A reference table for phases
}
\description{
create a list of phases in any project
}
\examples{
phases <- c("research", "drafting", "editing", "design", "print", "events")
phases <- create_phases(phases)
}
