% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-project-teams.R
\name{create_project_team}
\alias{create_project_team}
\title{Assigns staff roles on a project}
\usage{
create_project_team(project_teams)
}
\arguments{
\item{project_teams}{Data frame that lists project, role, staff, assigned_capacity}
}
\value{
A reference table project teams and roles
}
\description{
Assigns staff roles on a project
}
\examples{
n <- 3
projects <- LETTERS[1:n]
roles <- c("lead", "researcher", "editor", "design")
project_teams <- expand.grid(project = projects, role = roles)
staff <- c('Shelby', 'Luis', 'Taishawn', 'Samantha', 'Taylor')
project_teams$staff <- sample(staff, size = nrow(project_teams), replace = TRUE)
project_teams$assigned_capacity <- sample(c(25,25,75,100), size = nrow(project_teams), replace = TRUE)
project_teams <- create_project_team(project_teams)
}
