% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-projects.R
\name{create_projects}
\alias{create_projects}
\title{create a list of employees that are to be assigned to projects}
\usage{
create_projects(projects, probability = 1, start, end)
}
\arguments{
\item{projects}{Names of projects}

\item{probability}{Probability that the project will go ahead}

\item{start}{Expected start date of project}

\item{end}{Expected end date of the project}
}
\value{
A reference table for projects
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
library(lubridate)
n <- 3
projects <- LETTERS[1:n]
probability <- sample(c(50, 75, 100), size = n, replace = TRUE)
start <- sample(seq(as.Date(today() + 60), as.Date(today() + 365), by='day'), size = n, replace = FALSE) 
end <- start + sample(c(30,40,50,60), size = n, replace = TRUE)
projects <- create_projects(projects, probability, start, end)
}
