% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-responsibilities.R
\name{create_responsibilities}
\alias{create_responsibilities}
\title{creates a span responbsibilities of of roles across project phases}
\usage{
create_responsibilities(roles, phases, responsibilities)
}
\arguments{
\item{roles}{List of roles in any project team in order of responsibility}

\item{phases}{List of roles in any project team in order of responsibility}

\item{responsibilities}{a binary matrix of if role i is in volved in phase j}
}
\value{
A reference table for responsibilities
}
\description{
creates a span responbsibilities of of roles across project phases
}
\examples{
roles <- c("lead", "researcher", "editor", "design")
phases <- c("research", "drafting", "editing", "design", "print", "events")
responsibilities <- rbind(lead = rep(1, length(phases)), researcher = c(1,1,1,0,0,0), editor = c(0,0,1,0,0,0), design = c(0,0,0,1,1,0))
responsibilities <- create_responsibilities(roles, phases, responsibilities)
}
