% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-roles.R
\name{create_roles}
\alias{create_roles}
\title{create a list of roles in any project team}
\usage{
create_roles(roles)
}
\arguments{
\item{roles}{List of roles in any project team in order of responsibility}
}
\value{
A reference table for roles
}
\description{
create a list of roles in any project team
}
\examples{
roles <- c("lead", "researcher", "editor", "design")
roles <- create_roles(roles)
}
