% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-time-estimates.R
\name{create_time_estimates}
\alias{create_time_estimates}
\title{creates time estimates for each phases of each project}
\usage{
create_time_estimates(projects, phases, days_from_end)
}
\arguments{
\item{projects}{List of roles in any project team in order of responsibility}

\item{phases}{List of roles in any project team in order of responsibility}

\item{days_from_end}{Time estimates of how long each phase will take in relation to project end, negative = phase will occur before project end, positive = phase will occur after project end}
}
\value{
A reference table for time estimates of each phase of each project
}
\description{
creates time estimates for each phases of each project
}
\examples{
n <- 3
projects <- LETTERS[1:n]
phases <- c("research", "drafting", "editing", "design", "print", "events")
time_estimates <- rbind(A = c(-40,-10,-10,-10,-10,10), B = c(-10,-10,-10,0,0,0), C = c(0,0,0,-10,-10,10))
time_estimates <- create_time_estimates(projects, phases, time_estimates)
}
