% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-utils.R
\name{.db.get}
\alias{.db.get}
\title{Retrieve project information}
\usage{
.db.get(db, folder = "./db/")
}
\arguments{
\item{db}{Project table to retrieve}

\item{folder}{Location ofProject Tables}
}
\description{
Retrieve project information
}
