% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-utils.R
\name{.db.save}
\alias{.db.save}
\title{Save project information}
\usage{
.db.save(df, folder = "./db/")
}
\arguments{
\item{df}{Project table to save}

\item{folder}{Location to save Project Tables}
}
\description{
Save project information
}
