% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-set_full_schedule.R
\name{.get_assignment_schedule}
\alias{.get_assignment_schedule}
\title{pad out daily list of assignments for each staff}
\usage{
.get_assignment_schedule(schedule, wp)
}
\arguments{
\item{schedule}{A working schedule}

\item{wp}{Complete workplan object}
}
\value{
A reference table for daily projects schedule
}
\description{
pad out daily list of assignments for each staff
}
\keyword{internal}
