% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-set_full_schedule.R
\name{.get_pad_schedule}
\alias{.get_pad_schedule}
\title{pad out daily list of assignments for each staff}
\usage{
.get_pad_schedule(schedule)
}
\arguments{
\item{schedule}{A working schedule}
}
\value{
A reference table for daily projects schedule
}
\description{
pad out daily list of assignments for each staff
}
\keyword{internal}
