% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-set_full_schedule.R
\name{.get_schedule}
\alias{.get_schedule}
\title{create a daily list of assignments for each staff}
\usage{
.get_schedule(wp)
}
\arguments{
\item{wp}{Complete workplan object}
}
\value{
A reference table for daily projects schedule
}
\description{
create a daily list of assignments for each staff
}
\keyword{internal}
