% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_project.R
\docType{data}
\name{example_project}
\alias{example_project}
\title{Example project setup}
\format{A list with the follwoing data frames:
\describe{
  \item{resources}{price, in US dollars}
  \item{projects}{Lists of projects}
  \item{roles}{Roles within a project}
  \item{phases}{Phases of a project}
  \item{time_estimates}{Time required for different phases}
  \item{responsibilities}{A reference table for responsibilities}
  \item{project_teams}{Assigned project teams}
}}
\usage{
example_project
}
\description{
A dataset containing a premade project
}
\keyword{datasets}
