% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-workplan_io.R
\name{export_workplan}
\alias{export_workplan}
\title{Create Excel file for project inputs}
\usage{
export_workplan(wp, excel_file_name = "my_project.xlsx")
}
\arguments{
\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
\examples{
library(workplanr)
}
