% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-extract-plans.R
\name{extract_staff_schedule}
\alias{extract_staff_schedule}
\title{Add staff assignments to schedulke}
\usage{
extract_staff_schedule(staff, workplan)
}
\arguments{
\item{staff}{The name of the database to create}

\item{workplan}{schedule}
}
\description{
Add staff assignments to schedulke
}
