% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get-schedule.R
\name{factor_leave_in_work_allocation}
\alias{factor_leave_in_work_allocation}
\title{Re-calculate staff_contributions based on time in office}
\usage{
factor_leave_in_work_allocation(tmp)
}
\arguments{
\item{tmp}{schedule}
}
\description{
Re-calculate staff_contributions based on time in office
}
\keyword{internal}
