% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{full_schedule-class}
\alias{full_schedule-class}
\alias{full_sched}
\title{Base Class for workplan full_schedule}
\description{
Base Class for workplan full_schedule
}
\section{Slots}{

\describe{
\item{\code{projects}}{List of projects}

\item{\code{phases}}{List of phases in any project in order of execution}

\item{\code{date}}{Day wourk occurs on}

\item{\code{staff}}{Assigned staff to each [project, phase] combination}

\item{\code{assigned_capacity}}{Amount of time each staff is expected to dedicate to each [project, phase]}

\item{\code{capacity}}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}

\item{\code{public_holiday}}{= "character"}

\item{\code{out_of_office}}{= "character"}

\item{\code{project_duration}}{= "numeric"}

\item{\code{num_holidays}}{= "numeric"}

\item{\code{holiday_expansion_factor}}{= "numeric"}

\item{\code{num_out_of_office}}{= "numeric"}

\item{\code{leave_expansion_factor}}{= "numeric"}

\item{\code{leave_adjusted_workload}}{= "numeric"}
}}

\seealso{
Other classes: \code{\link{leave-class}},
  \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
