% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-get-daily-plan.R, R/backcast-schedule.R
\name{get.daily.plan}
\alias{get.daily.plan}
\title{Creates a randomised project management problem}
\usage{
get.daily.plan(resources, projects, time.estimates, public.holidays,
  project.teams, responsibilities, leave)

get.daily.plan(resources, projects, time.estimates, public.holidays,
  project.teams, responsibilities, leave)
}
\arguments{
\item{nothing}{}

\item{nothing}{}
}
\description{
Creates a randomised project management problem

Creates a randomised project management problem
}
