% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-headcount-table.R
\name{get.headcount.table}
\alias{get.headcount.table}
\title{ggplot colours}
\usage{
get.headcount.table(plan, filename = "./graphs/headcount-table.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
