% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-project-resource-allocation.R
\name{get.project.resource.allocation}
\alias{get.project.resource.allocation}
\title{ggplot colours}
\usage{
get.project.resource.allocation(plan,
  filename = "./graphs/resource-allocation-waterfall.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
