% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-project-resource-allocation-tile-plot.R
\name{get.project.resource.allocation.tile.plot}
\alias{get.project.resource.allocation.tile.plot}
\title{ggplot colours}
\usage{
get.project.resource.allocation.tile.plot(plan,
  filename = "./graphs/all-resource-tile-plot.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
