% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-project-teams.R
\name{get.project.teams}
\alias{get.project.teams}
\title{ggplot colours}
\usage{
get.project.teams(plan, filename = "./graphs/project-team-table.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
