% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-release-gantt-chart.R
\name{get.release.gannt.chart}
\alias{get.release.gannt.chart}
\title{ggplot colours}
\usage{
get.release.gannt.chart(plan,
  filename = "./graphs/release-gantt-chart.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
