% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-release-table.R
\name{get.release.table}
\alias{get.release.table}
\title{ggplot colours}
\usage{
get.release.table(plan, filename = "./graphs/release-table.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
