% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-team-capacity-timeline.R
\name{get.team.capacity.timeline}
\alias{get.team.capacity.timeline}
\title{ggplot colours}
\usage{
get.team.capacity.timeline(plan,
  filename = "./graphs/all-team-capacity-bar-plot.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
