% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-daily-plan.R
\name{get_daily_plan}
\alias{get_daily_plan}
\title{create a daily list of assignments for each staff}
\usage{
get_daily_plan(resources, projects, phases, time_estimates, project_teams,
  responsibilities)
}
\arguments{
\item{resources}{Names of projects}

\item{projects}{Probability that the project will go ahead}

\item{phases}{Expected start date of project}

\item{time_estimates}{Expected end date of the project}

\item{project_teams}{Expected end date of the project}

\item{responsibilites}{Expected end date of the project}
}
\value{
A reference table for projects
}
\description{
create a daily list of assignments for each staff
}
\examples{
data("resources")
data("projects")
data("phases")
data("roles")
data("responsibilities")
data("time_estimates")
data("project_teams")
get_daily_plan(esources, projects, phases, time_estimates, project_teams, responsibilities)
}
