% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_out_of_office}
\alias{get_out_of_office}
\title{Get out_of_office table}
\usage{
get_out_of_office(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Get out_of_office table
}
\keyword{internal}
