% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get-schedule.R
\name{get_phases}
\alias{get_phases}
\title{create a list of employees that are to be assigned to projects}
\usage{
get_phases(db_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
create a list of employees that are to be assigned to projects
}
\keyword{internal}
