% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{get_project_dependencies}
\alias{get_project_dependencies}
\title{Initialise schedule table}
\usage{
get_project_dependencies(workplan)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Initialise schedule table
}
