% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_project_phases}
\alias{get_project_phases}
\title{Get project_phases table}
\usage{
get_project_phases(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Get project_phases table
}
\keyword{internal}
