% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R,
%   R/11-get_project_teams.R
\name{get_project_teams}
\alias{get_project_teams}
\title{Add staff assignments to schedulke}
\usage{
get_project_teams(project, workplan)

get_project_teams(project, workplan)
}
\arguments{
\item{workplan}{schedule}

\item{workplan}{schedule}
}
\description{
Add staff assignments to schedulke

Add staff assignments to schedulke
}
\keyword{internal}
