% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_project_unassignments}
\alias{get_project_unassignments}
\title{Get project_assignments table}
\usage{
get_project_unassignments(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Get project_assignments table
}
\keyword{internal}
