% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_projects}
\alias{get_projects}
\title{Get projects table}
\usage{
get_projects(db_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Get projects table
}
\keyword{internal}
