% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_public_holidays}
\alias{get_public_holidays}
\title{get public_holidays table}
\usage{
get_public_holidays(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
get public_holidays table
}
\keyword{internal}
