% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-schedule_plotting.R
\name{get_related_projects}
\alias{get_related_projects}
\title{Get related projects}
\usage{
get_related_projects(project, schedule)
}
\arguments{
\item{project}{Project of interest}

\item{schedule}{Schedule}
}
\value{
staff_schedule if script exectues completely
}
\description{
Get related projects
}
\examples{
library(workplanr)
}
