% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{get_release_schedule}
\alias{get_release_schedule}
\title{Add staff assignments to schedulke}
\usage{
get_release_schedule(workplan)
}
\arguments{
\item{workplan}{schedule}
}
\description{
Add staff assignments to schedulke
}
\keyword{internal}
