% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-responsibilities.R
\name{get_responsibilities}
\alias{get_responsibilities}
\title{Retrieve responsibilities}
\usage{
get_responsibilities()
}
\description{
Retrieve responsibilities
}
\examples{
get_responsibilities()
}
