% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Create full schedule from database}
\usage{
get_schedule(workplan)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Create full schedule from database
}
