% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_staff}
\alias{get_staff}
\title{Get staff table}
\usage{
get_staff(db_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Get staff table
}
\keyword{internal}
