% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_time_estimates}
\alias{get_time_estimates}
\title{gt time_estimates table}
\usage{
get_time_estimates(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
gt time_estimates table
}
\keyword{internal}
