% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_workplan.R
\name{get_workplan}
\alias{get_workplan}
\title{create a list of employees that are to be assigned to projects}
\usage{
get_workplan(staff, staff_capacity, projects, project_probability,
  project_start, project_end, project_phases, project_time_estimates,
  staff_on_leave, leave_start, leave_end, leave_description,
  public_holidays_date, public_holidays_name,
  staff_project_assignment_capacity)
}
\arguments{
\item{staff}{Names of staff members}

\item{staff_capacity}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}

\item{projects}{Names of projects}

\item{project_probability}{Probability that the project will go ahead}

\item{project_start}{Expected start date of project}

\item{project_end}{Expected end date of the project}

\item{project_phases}{List of phases in any project in order of execution}

\item{project_time_estimates}{Time estimates of how long each phase will take in relation}

\item{staff_on_leave}{Names of staff that are going to be out of the office}

\item{leave_start}{Starting date for leave}

\item{leave_end}{Ending date for leave}

\item{leave_description}{Type of leave, can be user defined but recommend "leave" or "work trip"}

\item{public_holidays_date}{A data frame of dates of public holidays}

\item{public_holidays_name}{A data frame of names of public holidays}

\item{staff_project_assignment_capacity}{Amount of time each staff is expected to dedicate to each [project, phase]}
}
\value{
A reference table for staff
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
## ----library, include = TRUE, results='hide', message=FALSE, warning=FALSE----
library(workplanr)
## ----resources, include = TRUE, results='hide', message=FALSE, warning=FALSE----
staff <- c("Shelby", "Luis", "Taishawn", "Samantha", "Taylor", "unassigned")
staff_capacity <- c(40,60,100,100,100, 100)
## ----projects, include = TRUE, results='hide', message=FALSE, warning=FALSE----
projects <- LETTERS[1:3]
project_probability <- c(50, 100, 100)
project_start <- as.Date(c("2019-01-25", "2019-05-17", "2019-06-27"))
project_end <- as.Date(c("2019-06-03", "2019-06-16", "2019-09-27"))
project_phases <- c("research", "drafting", "editing", "design", "print", "events")
project_time_estimates  <- c(c(-10,-10,-10,-10,-10,-10), c(-10,-10,-10,-10,-10,-10), c(-10,-10,-10,10,10,10)) 
## ----leave, include = TRUE, results='hide', message=FALSE, warning=FALSE----
staff_on_leave <- c("Luis", "Samantha")
leave_start <-  as.Date(c("2019-07-23", "2019-05-16"))
leave_end <- leave_start + c(20, 25)
leave_description <- c("leave", "work trip")
## ----holidays, include = TRUE, results='hide', message=FALSE, warning=FALSE----
url <- "https://data.gov.au/data/dataset/b1bc6077-dadd-4f61-9f8c-002ab2cdff10/resource/31eec35e-1de6-4f04-9703-9be1d43d405b/download/australian_public_holidays_2019.csv"
public_holidays <- utils::read.csv(url, stringsAsFactors = FALSE)
names(public_holidays) <- tolower(names(public_holidays))
public_holidays$date <- as.Date(lubridate::ymd(public_holidays$date))
public_holidays <- public_holidays \%>\% filter(jurisdiction == "nsw") \%>\%
  select(date, holiday.name) \%>\% rename(name = holiday.name)
public_holidays_date <- public_holidays$date
public_holidays_name = public_holidays$name
## ----assignments, include = TRUE, results='hide', message=FALSE, warning=FALSE----
staff_project_assignment_capacity <- sample(c(0,25,50,75,100), size = length(projects)*length(project_phases)*length(staff), replace = TRUE)
## ----get_workplan, include = TRUE, results='hide', message=FALSE, warning=FALSE----
wp <- get_workplan(staff = staff, 
                   staff_capacity = staff_capacity,
                   projects = projects, 
                   project_probability =  project_probability, 
                   project_start = project_start, 
                   project_end = project_end, 
                   project_phases = project_phases, 
                   project_time_estimates = project_time_estimates, 
                   staff_on_leave = staff_on_leave, 
                   leave_start = leave_start, 
                   leave_end = leave_end, 
                   leave_description = leave_description, 
                   public_holidays_date = public_holidays_date,
                   public_holidays_name = public_holidays_name,
                   staff_project_assignment_capacity = staff_project_assignment_capacity)
print(wp)

}
