% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-create_workplan_db.R
\name{init_db}
\alias{init_db}
\title{create a list of employees that are to be assigned to projects}
\usage{
init_db(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
create a list of employees that are to be assigned to projects
}
\keyword{interal}
