% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get-schedule.R
\name{init_schedule}
\alias{init_schedule}
\title{Initialise schedule table}
\usage{
init_schedule(db_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Initialise schedule table
}
\keyword{internal}
