% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{leave-class}
\alias{leave-class}
\alias{leav}
\title{Base Class for workplan roles}
\description{
Base Class for workplan roles
}
\section{Slots}{

\describe{
\item{\code{staff_on_leave}}{Names of staff that are going to be out of the office}

\item{\code{leave_start}}{Starting date for leave}

\item{\code{leave_end}}{Ending date for leave}

\item{\code{leave_description}}{Type of leave, can be user defined but recommend "leave" or "work trip"}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{responsibilities-class}},
  \code{\link{roles-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
