% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-project-inputs.R
\name{load.project.inputs}
\alias{load.project.inputs}
\title{ggplot colours}
\usage{
load.project.inputs(filename = "./data/base-input.xlsx")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
