% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-normalise_inputs.R
\name{normalise_inputs}
\alias{normalise_inputs}
\title{Populate workplanr database with data}
\usage{
normalise_inputs(staff, projects, calendar, project_phases, project_roles,
  out_of_office, public_holidays, time_estimates, roles_responsibilities)
}
\arguments{
\item{staff}{Names of staff members}

\item{projects}{Names of projects}

\item{project_phases}{List of phases in any project in order of execution}

\item{out_of_office}{Names of staff that are going to be out of the office}

\item{public_holidays}{A data frame of dates of public holidays}

\item{time_estimates}{Time estimates of how long each phase will take in relation}

\item{roles_responsibilities}{Data frame for responsibilites of roles across project phases}
}
\description{
Populate workplanr database with data
}
\keyword{internal}
