% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{out_of_office-class}
\alias{out_of_office-class}
\alias{workplanr_out_of_office}
\title{Base Class for workplan roles}
\description{
Base Class for workplan roles
}
\section{Slots}{

\describe{
\item{\code{id_out_of_office}}{Names of staff that are going to be out of the office}

\item{\code{staff_name}}{Names of staff that are going to be out of the office}

\item{\code{out_of_office_start}}{Starting date for leave}

\item{\code{out_of_office_end}}{Ending date for leave}

\item{\code{work_related}}{Type of leave, can be user defined but recommend "leave" or "work trip"}
}}

\seealso{
Other classes: \code{\link{project_assignments-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{schedule-class}}, \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
