% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{out_of_office}
\alias{out_of_office}
\title{A dataset of periods where staff are out of the office}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{staff_name}{Staff who will be out of the office.}
  \item{out_of_office_start}{Date period starts.}
  \item{out_of_office_end}{Date of period end.}
  \item{work_related}{Whether out of office period is for work (TRUE) or vacation (FALSE)}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
out_of_office
}
\description{
A dataset containing periods where staff are out of the office
}
\keyword{datasets}
