% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{phases-class}
\alias{phases-class}
\alias{phas}
\title{Base Class for workplan phases}
\description{
Base Class for workplan phases
}
\section{Slots}{

\describe{
\item{\code{phases}}{List of phases in any project in order of execution}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
\keyword{internal}
