% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{plot,project_dependencies-method}
\alias{plot,project_dependencies-method}
\title{Coerce Object project_dependencies to a ggplot}
\usage{
\S4method{plot}{project_dependencies}(x)
}
\arguments{
\item{x}{A \code{project_dependencies} object.}
}
\description{
Coerce Object team_schedule to ggplot, avoiding using the "slot" notation.
}
