% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{plot,release_schedule-method}
\alias{plot,release_schedule-method}
\title{Coerce Object team_schedule to a ggplot}
\usage{
\S4method{plot}{release_schedule}(x)
}
\arguments{
\item{x}{A \code{team_schedule} object.}
}
\description{
Coerce Object team_schedule to ggplot, avoiding using the "slot" notation.
}
