% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{plot,staff_schedule-method}
\alias{plot,staff_schedule-method}
\title{Coerce Object staff_schedule to a ggplot}
\usage{
\S4method{plot}{staff_schedule}(x)
}
\arguments{
\item{x}{A \code{staff_schedule} object.}
}
\description{
Coerce Object full_schedule to ggplot, avoiding using the "slot" notation.
}
