% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{methods}
\name{plot,team_schedule-method}
\alias{plot,team_schedule-method}
\title{Coerce Object team_schedule to a ggplot}
\usage{
\S4method{plot}{team_schedule}(x)
}
\arguments{
\item{x}{A \code{team_schedule} object.}
}
\description{
Coerce Object team_schedule to ggplot, avoiding using the "slot" notation.
}
\examples{
library(workplanr)
wp <- build_sample_workplan()
plot(wp@staff_schedule)
}
