% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-plot-staff-calendar.R
\name{plot.staff.calendar}
\alias{plot.staff.calendar}
\title{ggplot colours}
\usage{
\method{plot}{staff.calendar}(plan, leave, public.holidays)
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
