% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-plot-team-workload.R
\name{plot.team.workload}
\alias{plot.team.workload}
\title{ggplot colours}
\usage{
\method{plot}{team.workload}(plan,
  filename = "./graphs/all-team-capacity-bar-plot.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
