% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-plot_daily_staff_plan.R
\name{plot_daily_staff_plan}
\alias{plot_daily_staff_plan}
\title{Plot a staff calendar}
\usage{
plot_daily_staff_plan(daily_plan)
}
\arguments{
\item{daily_plan}{Daily work plan of staff}
}
\value{
Calculated workload of staff per date
}
\description{
Plot a staff calendar
}
