% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-schedule_plotting.R
\name{plot_phase_days}
\alias{plot_phase_days}
\title{Plot staff schedule}
\usage{
plot_phase_days(workplan)
}
\arguments{
\item{schedule}{Schedule}
}
\value{
staff_schedule if script exectues completely
}
\description{
Plot staff schedule
}
\examples{
library(workplanr)
}
