% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-schedule_plotting.R
\name{plot_project_calendar}
\alias{plot_project_calendar}
\title{Plot project calendar}
\usage{
plot_project_calendar(project, schedule)
}
\arguments{
\item{project, }{Scheudle}

\item{schedule}{Scheudle}
}
\value{
staff_schedule if script exectues completely
}
\description{
Plot project calendar
}
\examples{
library(workplanr)
}
