% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-schedule_plotting.R
\name{plot_project_interactions}
\alias{plot_project_interactions}
\title{Plot project and related projects network}
\usage{
plot_project_interactions(project, schedule)
}
\arguments{
\item{project}{Project of interest}

\item{schedule}{Schedule}
}
\value{
staff_schedule if script exectues completely
}
\description{
Plot project and related projects network
}
\examples{
library(workplanr)
}
