% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plotting.R
\name{plot_staff_schedule}
\alias{plot_staff_schedule}
\title{create a list of employees that are to be assigned to projects}
\usage{
plot_staff_schedule(tmp)
}
\arguments{
\item{tmp}{Schedule}
}
\value{
staff_schedule if script exectues completely
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
library(workplanr)
}
