% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-schedule_plotting.R
\name{plot_staff_timeline}
\alias{plot_staff_timeline}
\title{Plot staff timeleine}
\usage{
plot_staff_timeline(staff, schedule)
}
\arguments{
\item{staff, }{Scheudle}

\item{schedule}{Scheudle}
}
\value{
staff_schedule if script exectues completely
}
\description{
Plot staff timeleine
}
\examples{
library(workplanr)
}
