% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-team-daily-workload.R
\name{plot_team_daily_workload}
\alias{plot_team_daily_workload}
\title{Plot a team workload calendar}
\usage{
plot_team_daily_workload(daily_plan)
}
\arguments{
\item{daily_plan}{Daily work plan of staff}
}
\value{
ggplot of workload of team per date
}
\description{
Plot a team workload calendar
}
