% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{project_assignments-class}
\alias{project_assignments-class}
\alias{workplanr_project_assignments}
\title{Base Class for workplan project_assignments}
\description{
Base Class for workplan project_assignments
}
\section{Slots}{

\describe{
\item{\code{project_name}}{List of projects}

\item{\code{project_role_name}}{Roles of projects}

\item{\code{project_phase_name}}{Phases}

\item{\code{staff_name}}{Assigned staff to each [project, role] combination (needs to be at least length(project) x length(roles) in length)}

\item{\code{staff_contribution}}{Amount of time each staff is expected to dedicate to each [project, role] 
combination (needs to be at least length(project) x length(roles) in length)}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{release_schedule-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{schedule-class}}, \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
