% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{project_assignments}
\alias{project_assignments}
\title{Project Assignments}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{project_name}{Name of project}
  \item{project_role_name}{Project roles}
  \item{staff_name}{Names of staff}
  \item{staff_contribution}{Approximate amount of time staff will dedicate to performing that role. Needs to be the same scale as staff_capacity in data("staff")}
  ...
}}
\usage{
project_assignments
}
\description{
A dataset containing the staff assigned to roles on projects
}
\keyword{datasets}
