% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-input_checks.R
\name{project_name_check}
\alias{project_name_check}
\title{Populate workplanr database with data}
\usage{
project_name_check(project_name, workplan)
}
\arguments{
\item{project_name}{Names of staff members}

\item{workplan}{Workplan}
}
\description{
Populate workplanr database with data
}
\keyword{internal}
