% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{project_phases}
\alias{project_phases}
\title{Project phases}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{project_phase_name}{Phases of a project}
  ...
}}
\usage{
project_phases
}
\description{
A dataset containing the the project phases
}
\keyword{datasets}
