% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{project_roles-class}
\alias{project_roles-class}
\alias{workplanr_project_roles}
\title{Base Class for workplan roles}
\description{
Base Class for workplan roles
}
\section{Slots}{

\describe{
\item{\code{project_role_name}}{List of roles in any project team in order of responsibility}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_assignments-class}},
  \code{\link{project_dependencies-class}},
  \code{\link{project_phases-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{release_schedule-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{schedule-class}}, \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
