% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{project_roles}
\alias{project_roles}
\title{Project roles}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{project_role_name}{Project roles}
  ...
}}
\usage{
project_roles
}
\description{
A dataset containing the the project phases
}
\keyword{datasets}
