% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{project_teams-class}
\alias{project_teams-class}
\alias{team}
\title{Base Class for workplan project_teams}
\description{
Base Class for workplan project_teams
}
\section{Slots}{

\describe{
\item{\code{projects}}{List of projects}

\item{\code{phases}}{List of phases in any project in order of execution}

\item{\code{staff}}{Names of staff members}

\item{\code{assigned_capacity}}{Amount of time each staff is expected to dedicate to each [project, phase]}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
