% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{projects-class}
\alias{projects-class}
\alias{proj}
\title{Base Class for workplan projects}
\description{
Base Class for workplan projects
}
\section{Slots}{

\describe{
\item{\code{projects}}{Names of projects}

\item{\code{probability}}{Probability that the project will go ahead}

\item{\code{start}}{Expected start date of project}

\item{\code{end}}{Expected end date of the project}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
