% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{projects}
\alias{projects}
\title{Data frame of proejct information}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{project_name}{Project Name}
  \item{project_confirmed}{Whether the project is confirmed TRUE = confirmed, FALSE = potential work, allows project manager to plan for unconfirmed work}
  \item{project_start}{Date as character for project start}
  \item{project_end}{Date of end of project}
  ...
}}
\usage{
projects
}
\description{
A dataset of projects, whether they are confirmed, start and end dates
}
\keyword{datasets}
