% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-input_checks.R
\name{proper_capitalise}
\alias{proper_capitalise}
\title{Populate workplanr database with data}
\usage{
proper_capitalise(string)
}
\arguments{
\item{string}{Names of staff members}
}
\description{
Populate workplanr database with data
}
\keyword{internal}
