% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{public_holidays-class}
\alias{public_holidays-class}
\alias{holi}
\title{Base Class for workplan roles}
\description{
Base Class for workplan roles
}
\section{Slots}{

\describe{
\item{\code{date}}{Date of public holiday}

\item{\code{name}}{Name of public holiday}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{resources-class}},
  \code{\link{responsibilities-class}},
  \code{\link{roles-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
