% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{public_holidays}
\alias{public_holidays}
\title{Public Holidays}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{date}{The date of the public holiday}
  \item{holiday_name}{The name of the holiday}
  ...
}}
\usage{
public_holidays
}
\description{
A dataset containing the dates of public holidays in New South Wales, Australia in 2019.
}
\keyword{datasets}
