% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-workplanr-io.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Create Excel file for project inputs}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{tibble}{Read in a tibble?}

\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
\examples{
library(workplanr)
}
\keyword{internal}
