% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-assign_and_remove.R
\name{remove_staff}
\alias{remove_staff}
\title{Remove staff member from project phase}
\usage{
remove_staff(db_name, staff_name, project_name, project_role_name)
}
\arguments{
\item{db_name}{The name of the database to create}

\item{staff_name}{staff name}

\item{project_name}{project name}

\item{project_phase_name}{project_phase_name}
}
\description{
Remove staff member from project phase
}
