% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{resources-class}
\alias{resources-class}
\alias{resc}
\title{Base Class for workplan resources}
\description{
Base Class for workplan resources
}
\section{Slots}{

\describe{
\item{\code{staff}}{Names of staff members}

\item{\code{capacity}}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
\keyword{internal}
