% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{roles_responsibilities-class}
\alias{roles_responsibilities-class}
\alias{workplanr_roles_responsibilities}
\title{Base Class for workplan responsibilites}
\description{
Base Class for workplan responsibilites
}
\section{Slots}{

\describe{
\item{\code{project_role_name}}{List of roles in any project team in order of responsibility}

\item{\code{project_phase_name}}{List of roles in any project team in order of responsibility}

\item{\code{responsibility_span}}{a binary matrix of if role i is in volved in phase j}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_assignments-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{schedule-class}}, \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
