% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{roles_responsibilities}
\alias{roles_responsibilities}
\title{Roles and responsibilities}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{project_role_name}{Project Role}
  \item{project_phase_name}{Project phase}
  \item{responsibility_span}{Whether the role in involved in that phase of the project, 1 = yes, 0 = no.}
  ...
}}
\usage{
roles_responsibilities
}
\description{
A dataset containing the resposibilities across phases that each project role has
}
\keyword{datasets}
