% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{schedule-class}
\alias{schedule-class}
\alias{workplanr_schedule}
\title{Base Class for schedule project_assignments}
\description{
Base Class for schedule project_assignments
}
\section{Slots}{

\describe{
\item{\code{date}}{List of projects}

\item{\code{project_name}}{List of projects}

\item{\code{project_confirmed}}{project confirmed}

\item{\code{project_phase_name}}{Phases}

\item{\code{project_role_name}}{Roles of projects}

\item{\code{staff_name}}{Names of staff}

\item{\code{staff_capacity}}{Capacity of staff}

\item{\code{staff_contribution}}{Contibution of styaff to project}

\item{\code{holiday_name}}{names of public holidays
combination (needs to be at least length(project) x length(roles) in length)}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_assignments-class}},
  \code{\link{project_dependencies-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{release_schedule-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
\keyword{internal}
