% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-set_full_schedule.R
\name{set_full_schedule}
\alias{set_full_schedule}
\title{create a daily list of assignments for each staff}
\usage{
set_full_schedule(wp)
}
\arguments{
\item{wp}{Complete workplan object}
}
\value{
A reference table for daily projects schedule
}
\description{
create a daily list of assignments for each staff
}
\keyword{internal}
