% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-set_full_schedule.R
\name{set_holidays}
\alias{set_holidays}
\title{pad out daily list of assignments for each staff}
\usage{
set_holidays(schedule, tmp)
}
\arguments{
\item{schedule}{A working schedule}

\item{tmp}{Complete workplan object}
}
\value{
A reference table for daily projects schedule
}
\description{
pad out daily list of assignments for each staff
}
\keyword{internal}
