% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-leave.R
\name{set_leave}
\alias{set_leave}
\title{create a list of phases in any project}
\usage{
set_leave(staff_on_leave, leave_start, leave_end, leave_description)
}
\arguments{
\item{phases}{List of phases in any project in order of execution}
}
\value{
A reference table for phases
}
\description{
create a list of phases in any project
}
\keyword{internal}
