% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-set_full_schedule.R
\name{set_project_phase_dates}
\alias{set_project_phase_dates}
\title{create a daily list of assignments for each staff}
\usage{
set_project_phase_dates(tmp)
}
\arguments{
\item{tmp}{Complete workplan object}
}
\value{
A reference table for daily projects schedule
}
\description{
create a daily list of assignments for each staff
}
\keyword{internal}
