% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-set_project_teams.R
\name{set_project_team}
\alias{set_project_team}
\title{Assigns staff roles on a project}
\usage{
set_project_team(projects, roles, assigned_staff = NA,
  assigned_capacity = NA)
}
\arguments{
\item{projects}{List of projects}

\item{roles}{Roles of projects}

\item{assigned_capacity}{Amount of time each staff is expected to dedicate to each [project, role] 
combination (needs to be at least length(project) x length(roles) in length)}

\item{staff}{Assigned staff to each [project, role] combination (needs to be at least length(project) x length(roles) in length)}
}
\value{
A reference table project teams and roles
}
\description{
Assigns staff roles on a project
}
\keyword{internal}
