% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-projects.R
\name{set_projects}
\alias{set_projects}
\title{create a list of employees that are to be assigned to projects}
\usage{
set_projects(projects, probability = 1, start, end)
}
\arguments{
\item{projects}{Names of projects}

\item{probability}{Probability that the project will go ahead}

\item{start}{Expected start date of project}

\item{end}{Expected end date of the project}
}
\value{
A reference table for projects
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
projects <- LETTERS[1:3]
probability <- c(50, 100, 100)
start <- as.Date(c("2019-07-25", "2019-05-17", "2019-09-27")) 
end <- as.Date(c("2019-09-03", "2019-06-16", "2019-10-27"))
projects <- set_projects(projects, probability, start, end)
}
