% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-resources.R
\name{set_resources}
\alias{set_resources}
\title{create a list of employees that are to be assigned to projects}
\usage{
set_resources(staff, capacity)
}
\arguments{
\item{staff}{Names of staff members}

\item{capacity}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}
}
\value{
A reference table for staff
}
\description{
create a list of employees that are to be assigned to projects
}
