% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-responsibilities.R
\name{set_responsibilities}
\alias{set_responsibilities}
\title{creates a span responbsibilities of of roles across project phases}
\usage{
set_responsibilities(roles, phases, responsibilities = 1)
}
\arguments{
\item{roles}{List of roles in any project team in order of responsibility}

\item{phases}{List of roles in any project team in order of responsibility}

\item{responsibilities}{a binary matrix of if role i is in volved in phase j}
}
\value{
A reference table for responsibilities
}
\description{
creates a span responbsibilities of of roles across project phases
}
