% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-roles.R
\name{set_roles}
\alias{set_roles}
\title{create a list of roles in any project team}
\usage{
set_roles(roles)
}
\arguments{
\item{roles}{List of roles in any project team in order of responsibility}
}
\value{
A reference table for roles
}
\description{
create a list of roles in any project team
}
\keyword{internal}
