% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-set_staff_schedule.R
\name{set_staff_schedule}
\alias{set_staff_schedule}
\title{Create a daily list of assignments for each staff}
\usage{
set_staff_schedule(wp)
}
\arguments{
\item{daily_plan}{Daily work plan of staff}
}
\value{
Calculated workload of staff per date
}
\description{
Create a daily list of assignments for each staff
}
\keyword{internal}
