% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-set_team_schedule.R
\name{set_team_schedule}
\alias{set_team_schedule}
\title{Calculate team workload calendar}
\usage{
set_team_schedule(wp)
}
\arguments{
\item{daily_plan}{Daily work plan of staff}
}
\value{
Calculated workload of team per date
}
\description{
Calculate team workload calendar
}
\keyword{internal}
