% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-time_estimates.R
\name{set_time_estimates}
\alias{set_time_estimates}
\title{creates time estimates for each phases of each project}
\usage{
set_time_estimates(projects, phases, time_estimates = 10)
}
\arguments{
\item{projects}{List of roles in any project team in order of responsibility}

\item{phases}{List of roles in any project team in order of responsibility}

\item{time_estimates}{Time estimates of how long each phase will take in relation to project end, negative = phase will occur before project end, positive = phase will occur after project end}
}
\value{
A reference table for time estimates of each phase of each project
}
\description{
creates time estimates for each phases of each project
}
