% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-shift_projects.R
\name{shift_project}
\alias{shift_project}
\title{Assign staff member to project phase}
\usage{
shift_project(project, by = NULL, new_start_date = NULL,
  new_end_date = NULL, workplan)
}
\arguments{
\item{project}{The name of the database to create}

\item{by}{staff name}

\item{new_start_date}{project name}

\item{new_end_date}{project_phase_name}

\item{workplan}{wp}
}
\description{
Assign staff member to project phase
}
