% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{staff-class}
\alias{staff-class}
\alias{workplanr_staff}
\title{Base Class for workplan resources}
\description{
Base Class for workplan resources
}
\section{Slots}{

\describe{
\item{\code{staff_name}}{Names of staff members}

\item{\code{staff_capacity}}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_assignments-class}},
  \code{\link{project_dependencies-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{release_schedule-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{schedule-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
\keyword{internal}
