% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{staff}
\alias{staff}
\title{Exmaple list of staff}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{staff_name}{Names of staff}
  \item{staff_capacity}{Capacity, can be on any scale but easiest to use scale of 100 for full time equivalent, 
  40 for someone who works two days a week, etc.}
  ...
}}
\usage{
staff
}
\description{
A dataset staff_names and staff_capacity
}
\keyword{datasets}
