% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-input_checks.R
\name{staff_name_check}
\alias{staff_name_check}
\title{Populate workplanr database with data}
\usage{
staff_name_check(staff_name, workplan)
}
\arguments{
\item{workplan}{Workplan}

\item{staff}{Names of staff members}
}
\description{
Populate workplanr database with data
}
\keyword{internal}
