% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{staff_schedule-class}
\alias{staff_schedule-class}
\alias{staff_sched}
\title{Base Class for workplan staff_schedule}
\description{
Base Class for workplan staff_schedule
}
\section{Slots}{

\describe{
\item{\code{date}}{Day work occurs on}

\item{\code{staff}}{Assigned staff to each [project, role] combination}

\item{\code{projects}}{Names of projects}

\item{\code{workload}}{Amount of work assigned to staff member}

\item{\code{out_of_office}}{Whether a staff member is out of office on a date}

\item{\code{public_holiday}}{Whether a day is a public holiday}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{responsibilities-class}},
  \code{\link{roles-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
