% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-team-workload.R
\name{team.workload}
\alias{team.workload}
\title{ggplot colours}
\usage{
team.workload(plan, filename = "./graphs/all-team-capacity-bar-plot.png")
}
\arguments{
\item{p}{a ggplot}
}
\description{
ggplot colours
}
