% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{team_schedule-class}
\alias{team_schedule-class}
\alias{team_sched}
\title{Base Class for workplan team_schedule}
\description{
Base Class for workplan team_schedule
}
\section{Slots}{

\describe{
\item{\code{date}}{Day wourk occurs on}

\item{\code{workload}}{Amount of work assigned to staff member}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
\keyword{internal}
