% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{time_estimates-class}
\alias{time_estimates-class}
\alias{time}
\title{Base Class for workplan time_estimates}
\description{
Base Class for workplan time_estimates
}
\section{Slots}{

\describe{
\item{\code{projects}}{List of roles in any project team in order of responsibility}

\item{\code{phases}}{List of roles in any project team in order of responsibility}

\item{\code{time_estimates}}{Time estimates of how long each phase will take in relation 
to project end, negative = phase will occur before project end, positive = phase will occur after project end}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{responsibilities-class}},
  \code{\link{roles-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
