% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-data-workplan_inputs.R
\docType{data}
\name{time_estimates}
\alias{time_estimates}
\title{Time Estimates}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{project_name}{The project}
  \item{project_phase_name}{The pahse of the project}
  \item{time_estimate}{Amount of time that each phase is expected to be in each project. 
  A negative number means that the phase occurs before the end date of the project, 
  a positive number means it happens after the end date of a project, 
  and 0 means phase will not occur in the project}
  ...
}}
\usage{
time_estimates
}
\description{
A dataset containing the estimates of time for each pahse in each project.
}
\keyword{datasets}
