% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\docType{class}
\name{workplan-class}
\alias{workplan-class}
\alias{workplan}
\title{Base Class for workplan workplan}
\description{
Base Class for workplan workplan
}
\section{Slots}{

\describe{
\item{\code{resources}}{Object of class "resource"}

\item{\code{projects}}{Object of class "projects"}

\item{\code{phases}}{Object of class "phases",}

\item{\code{leave}}{Object of class  "leave",}

\item{\code{holidays}}{Object of class  "public_holidays",}

\item{\code{time_estimates}}{Object of class  "time_estimates",}

\item{\code{project_teams}}{Object of class  "project_teams"}

\item{\code{full_schedule}}{Object of class "full_schedule"}

\item{\code{staff_schedule}}{Object of class "staff_schedule"}

\item{\code{team_schedule}}{Object of class "team_schedule"}
}}

\seealso{
Other classes: \code{\link{full_schedule-class}},
  \code{\link{leave-class}}, \code{\link{phases-class}},
  \code{\link{project_teams-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{resources-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}}
}
\concept{classes}
\keyword{internal}
