% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get-schedule.R
\name{add_project_calendar}
\alias{add_project_calendar}
\title{Add calendar to schedule}
\usage{
add_project_calendar(tmp, db_name)
}
\arguments{
\item{tmp}{schedule}

\item{db_name}{The name of the database to create}
}
\description{
Add calendar to schedule
}
\keyword{internal}
