% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{add_staff_out_of_office}
\alias{add_staff_out_of_office}
\title{Add out of office times to schedule}
\usage{
add_staff_out_of_office(schedule, workplan)
}
\arguments{
\item{db_name}{The name of the database to create}

\item{tmp}{schedule}
}
\description{
Add out of office times to schedule
}
\keyword{internal}
