% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-schedule.R
\name{add_staff_out_of_office}
\alias{add_staff_out_of_office}
\title{Add staff assignments to schedulke}
\usage{
add_staff_out_of_office(schedule, workplan)
}
\arguments{
\item{schedule}{The name of the database to create}

\item{workplan}{schedule}
}
\description{
Add staff assignments to schedulke
}
\keyword{internal}
