% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{as.data.frame,team_schedule-method}
\alias{as.data.frame,team_schedule-method}
\title{Coerce Object project_teams to a data frame}
\usage{
\S4method{as.data.frame}{team_schedule}(x)
}
\arguments{
\item{x}{A \code{project_teams} object.}
}
\description{
Coerce Object project_teams to a data frame, avoiding using the "slot" notation.
}
\keyword{internal}
