% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-create_new_workplan_db.R
\name{create_new_workplan_db}
\alias{create_new_workplan_db}
\title{Populate workplanr database with data}
\usage{
create_new_workplan_db(staff, projects, calendar, project_phases,
  out_of_office, public_holidays, time_estimates,
  staff_name_for_unassigned_work = "unassigned",
  db_name = "my_workplan.sqlite")
}
\arguments{
\item{staff}{Names of staff members}

\item{projects}{Names of projects}

\item{calendar}{Dates of work plan}

\item{project_phases}{List of phases in any project in order of execution}

\item{out_of_office}{Names of staff that are going to be out of the office}

\item{public_holidays}{A data frame of dates of public holidays}

\item{time_estimates}{Time estimates of how long each phase will take in relation}

\item{staff_name_for_unassigned_work}{Dummy staff member to assign work to}

\item{db_name}{name of database to create}
}
\value{
TRUE if script exectues completely
}
\description{
Populate workplanr database with data
}
\examples{
library(workplanr)
}
