% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_workplan_classes.R
\name{create_project_file}
\alias{create_project_file}
\title{Create Excel file for project inputs}
\usage{
create_project_file(plan = utils::data("example_project", envir =
  environment()), excel_file_name = "my_project.xlsx")
}
\arguments{
\item{plan}{File name for project inputs}

\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
\examples{
create_project_file()
}
