% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-utils.R
\name{create_project_file}
\alias{create_project_file}
\title{Crete Excel file for project inputs}
\usage{
create_project_file(excel_file_name = "my_project.xlsx")
}
\arguments{
\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
