% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-workplan_io.R
\name{export_workplan}
\alias{export_workplan}
\title{Create Excel file for project inputs}
\usage{
export_workplan(wp, excel_file_name = "my_workplan.xlsx")
}
\arguments{
\item{wp}{A workplan object}

\item{excel_file_name}{File name for project inputs}
}
\description{
This function creates an excel file that can be used to create a new project
}
\examples{
library(workplanr)
wp <- build_sample_workplan()
## ----export_workplan, include = TRUE, results='hide', message=FALSE, warning=FALSE----
export_workplan(wp, excel_file_name = "my-workplan.xlsx")
## ----import_workplan, include = TRUE, results='hide', message=FALSE, warning=FALSE----
wp <- import_workplan(excel_file_name = "my-workplan.xlsx")
}
