% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-extract-plans.R
\name{extract_project_schedule}
\alias{extract_project_schedule}
\title{Add staff assignments to schedulke}
\usage{
extract_project_schedule(project, workplan)
}
\arguments{
\item{project}{The name of the database to create}

\item{workplan}{schedule}
}
\description{
Add staff assignments to schedulke
}
