% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-extract-plans.R
\name{extract_staff_schedule}
\alias{extract_staff_schedule}
\title{Initialise schedule table}
\usage{
extract_staff_schedule(staff, workplan)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Initialise schedule table
}
