% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_functions.R
\name{get_calendar}
\alias{get_calendar}
\title{get calendar table}
\usage{
get_calendar(db_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
get calendar table
}
\keyword{internal}
