% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-daily_plan.R
\name{get_daily_plan}
\alias{get_daily_plan}
\title{create a daily list of assignments for each staff}
\usage{
get_daily_plan(wp)
}
\arguments{
\item{plan}{Complete workplan object}
}
\value{
A reference table for daily projects schedule
}
\description{
create a daily list of assignments for each staff
}
