% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-daily-staff-plan.R
\name{get_staff_daily_workload}
\alias{get_staff_daily_workload}
\title{Create a daily list of assignments for each staff}
\usage{
get_staff_daily_workload(daily_plan)
}
\arguments{
\item{daily_plan}{Daily work plan of staff}
}
\value{
Calculated workload of staff per date
}
\description{
Create a daily list of assignments for each staff
}
