% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-team-daily-workload.R
\name{get_team_daily_workload}
\alias{get_team_daily_workload}
\title{Calculate team workload calendar}
\usage{
get_team_daily_workload(daily_plan)
}
\arguments{
\item{daily_plan}{Daily work plan of staff}
}
\value{
Calculated workload of team per date
}
\description{
Calculate team workload calendar
}
