% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get-schedule.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Create team schedule from full schedule}
\usage{
get_team_schedule(tmp)
}
\arguments{
\item{tmp}{Scheudle}
}
\value{
staff_schedule if script exectues completely
}
\description{
Create team schedule from full schedule
}
\keyword{internal}
