% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_time_estimates}
\alias{get_time_estimates}
\title{create a list of employees that are to be assigned to projects}
\usage{
get_time_estimates(db_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
create a list of employees that are to be assigned to projects
}
\keyword{internal}
