% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-create_workplan.R
\name{get_workplan}
\alias{get_workplan}
\title{create a list of employees that are to be assigned to projects}
\usage{
get_workplan(staff, capacity, projects, probability, start, end, phases,
  roles, responsibilities = 1, time_estimates = 10,
  assigned_staff = "unassigned", assigned_capacity = 100)
}
\arguments{
\item{staff}{Names of staff members}

\item{capacity}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}
}
\value{
A reference table for staff
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
staff <- c('Shelby', 'Luis', 'Taishawn', 'Samantha', 'Taylor')
capacity <- c(40,60,100,100,100)
projects <- LETTERS[1:3]
probability <- c(50, 100, 100)
start <- as.Date(c("2019-07-25", "2019-05-17", "2019-09-27")) 
end <- as.Date(c("2019-09-03", "2019-06-16", "2019-10-27"))
phases <- c("research", "drafting", "editing", "design", "print", "events")
roles <- c("lead", "researcher", "editor", "design")
wp <- get_workplan(staff, capacity, projects, probability, start, end, phases, roles)
}
