% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_workplan.R
\name{get_workplan}
\alias{get_workplan}
\title{create a list of employees that are to be assigned to projects}
\usage{
get_workplan(staff, staff_capacity, projects, project_probability,
  project_start, project_end, project_phases, project_roles,
  project_responsibilities = 1, project_time_estimates = 10,
  staff_on_leave, leave_start, leave_end, leave_description,
  public_holidays, staff_project_assignments = "unassigned",
  staff_project_assigned_capacity = 100)
}
\arguments{
\item{staff}{Names of staff members}

\item{staff_capacity}{Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week}

\item{projects}{Names of projects}

\item{project_probability}{Probability that the project will go ahead}

\item{project_start}{Expected start date of project}

\item{project_end}{Expected end date of the project}

\item{project_phases}{List of phases in any project in order of execution}

\item{project_roles}{List of roles in any project team in order of responsibility}

\item{project_responsibilities}{a binary matrix of if role i is in volved in phase j}

\item{project_time_estimates}{Time estimates of how long each phase will take in relation}

\item{staff_on_leave}{Names of staff that are going to be out of the office}

\item{leave_start}{Starting date for leave}

\item{leave_end}{Ending date for leave}

\item{leave_description}{Type of leave, can be user defined but recommend "leave" or "work trip"}

\item{public_holidays}{A data frame of dates and names of public holidays}

\item{staff_project_assignments}{Assigned staff to each [project, role] combination (needs to be at least length(project) x length(roles) in length)}

\item{staff_project_assigned_capacity}{Amount of time each staff is expected to dedicate to each [project, role] 
(needs to be at least length(project) x length(roles) in length)}
}
\value{
A reference table for staff
}
\description{
create a list of employees that are to be assigned to projects
}
\examples{
library(workplanr)
}
