% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-create_workplan_db.R
\name{init_db}
\alias{init_db}
\title{Initialise workplanr database}
\usage{
init_db(db_name = "my_workplan.sqlite")
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
Initialise workplanr database
}
\keyword{interal}
