% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{methods}
\name{plot,staff_schedule-method}
\alias{plot,staff_schedule-method}
\title{#' Coerce Object workplan to a list of data frames
#'
#' @description Coerce Object workplan to a list of data frames, avoiding using the "slot" notation.
#'
#' @param x A \code{workplan} object.
#' @export
setMethod("as.list", "workplan", definition = function(x){
  x = list(
    resources = as.data.frame(x@resources),
    projects = as.data.frame(x@projects),
    phases = as.data.frame(x@phases),
    roles = as.data.frame(x@roles),
    leave = as.data.frame(x@leave),
    holidays = as.data.frame(x@holidays),
    responsibilities = as.data.frame(x@responsibilities),
    time_estimates = as.data.frame(x@time_estimates),
    project_teams = as.data.frame(x@project_teams),
    full_schedule = as.data.frame(x@full_schedule),
    staff_schedule = as.data.frame(x@staff_schedule),
    team_schedule = as.data.frame(x@team_schedule))
  return(x)
})}
\usage{
\S4method{plot}{staff_schedule}(x)
}
\arguments{
\item{x}{A \code{staff_schedule} object.}
}
\description{
Coerce Object full_schedule to ggplot, avoiding using the "slot" notation.
}
\details{
#' Base Class for workplan full_schedule
#'
#' @slot projects List of projects
#' @slot phases List of phases in any project in order of execution
#' @slot date Day wourk occurs on
#' @slot role Roles of projects
#' @slot staff Assigned staff to each [project, role] combination 
#' @slot assigned_capacity Amount of time each staff is expected to dedicate to each [project, role] 
#' @slot capacity Number of units of work per staff, for example 100 for full time equivalents, 40 for staff who work only 2 days per week
#' @slot public_holiday = "character"
#' @slot out_of_office = "character" 
#' @slot project_duration = "numeric"
#' @slot num_holidays = "numeric" 
#' @slot holiday_expansion_factor = "numeric"
#' @slot num_out_of_office = "numeric" 
#' @slot leave_expansion_factor = "numeric"
#' @slot leave_adjusted_workload = "numeric"
#' @family classes
full_sched <- setClass("full_schedule", slots = c(date = "Date", project = "ordered", phase = "ordered", 
                                                  role = "ordered", staff = "character", assigned_capacity = "numeric",
                                                  capacity = "numeric", public_holiday = "character",
                                                  out_of_office = "character", project_duration = "numeric",
                                                  num_holidays = "numeric", holiday_expansion_factor = "numeric",
                                                  num_out_of_office = "numeric", leave_expansion_factor = "numeric",
                                                  leave_adjusted_workload = "numeric"))

#' Coerce Object full_schedule to a data frame
#'
#' @description Coerce Object full_schedule to a data frame, avoiding using the "slot" notation.
#'
#' @param x A \code{full_schedule} object.
setMethod("as.data.frame", "full_schedule", definition = function(x){
  x <- data.frame(date = x@date, project = x@project, phase = x@phase,
                  role = x@role, staff = x@staff, assigned_capacity = x@assigned_capacity, capacity = x@capacity,
                  public_holiday = x@public_holiday, out_of_office = x@out_of_office, project_duration = x@project_duration,
                  num_holidays = x@num_holidays, holiday_expansion_factor = x@holiday_expansion_factor,
                  num_out_of_office = x@num_out_of_office, leave_expansion_factor = x@leave_expansion_factor,
                  leave_adjusted_workload = x@leave_adjusted_workload)
  return(x)
})

#' Base Class for workplan staff_schedule
#'
#' @slot date Day work occurs on
#' @slot staff Assigned staff to each [project, role] combination 
#' @slot projects Names of projects
#' @slot workload Amount of work assigned to staff member 
#' @slot out_of_office Whether a staff member is out of office on a date
#' @slot public_holiday Whether a day is a public holiday
#' @family classes
staff_sched <- setClass("staff_schedule", slots = c(date = "Date", staff = "character", project = "character", workload= "numeric",
                                                    out_of_office = "character", 
                                                    public_holiday = "character"))

#' Coerce Object staff_schedule to a data frame
#'
#' @description Coerce Object staff_schedule to a data frame, avoiding using the "slot" notation.
#'
#' @param x A \code{staff_schedule} object.
setMethod("as.data.frame", "staff_schedule", definition = function(x){
  x <- data.frame(date = x@date, staff = x@staff, project = x@project, workload = x@workload, 
                  out_of_office = x@out_of_office, public_holiday = x@public_holiday)
  return(x)
})

Coerce Object staff_schedule to a ggplot
}
