% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-schedule_plotting.R
\name{plot_staff_schedule}
\alias{plot_staff_schedule}
\title{Plot staff schedule}
\usage{
plot_staff_schedule(tmp)
}
\arguments{
\item{tmp}{Schedule}
}
\value{
staff_schedule if script exectues completely
}
\description{
Plot staff schedule
}
\examples{
library(workplanr)
}
