% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plotting.R
\name{plot_team_schedule}
\alias{plot_team_schedule}
\title{Plot team schedule}
\usage{
plot_team_schedule(tmp)
}
\arguments{
\item{tmp}{Scheudle}
}
\value{
staff_schedule if script exectues completely
}
\description{
Plot team schedule
}
\examples{
library(workplanr)
}
