% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{projects-class}
\alias{projects-class}
\alias{workplanr_projects}
\title{Base Class for workplan projects}
\description{
Base Class for workplan projects
}
\section{Slots}{

\describe{
\item{\code{project_name}}{Names of projects}

\item{\code{project_confirmed}}{Probability that the project will go ahead}

\item{\code{project_start}}{Expected start date of project}

\item{\code{project_end}}{Expected end date of the project}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_assignments-class}},
  \code{\link{project_dependencies-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{public_holidays-class}},
  \code{\link{release_schedule-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{schedule-class}}, \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{time_estimates-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
