% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-assign_and_remove.R
\name{remove_staff}
\alias{remove_staff}
\title{create a list of employees that are to be assigned to projects}
\usage{
remove_staff(db_name, staff_name, project_name, project_phase_name)
}
\arguments{
\item{db_name}{The name of the database to create}
}
\description{
create a list of employees that are to be assigned to projects
}
