% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-set_leave.R, R/12-set_full_schedule.R
\name{set_leave}
\alias{set_leave}
\title{Create an object of class leave}
\usage{
set_leave(schedule, tmp)

set_leave(schedule, tmp)
}
\arguments{
\item{schedule}{A working schedule}

\item{tmp}{Complete workplan object}

\item{staff_on_leave}{Names of staff that are going to be out of the office}

\item{leave_start}{Starting date for leave}

\item{leave_end}{Ending date for leave}

\item{leave_description}{Type of leave, can be user defined but recommend "leave" or "work trip"}
}
\value{
A reference table for staff leave

A reference table for daily projects schedule
}
\description{
Create an object of class leave

pad out daily list of assignments for each staff
}
\keyword{internal}
