% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-phases.R, R/06-leave.R, R/07-holidays.R
\name{set_phases}
\alias{set_phases}
\title{create a list of phases in any project}
\usage{
set_phases(phases)

set_phases(phases)

set_phases(phases)
}
\arguments{
\item{phases}{List of phases in any project in order of execution}

\item{phases}{List of phases in any project in order of execution}

\item{phases}{List of phases in any project in order of execution}
}
\value{
A reference table for phases

A reference table for phases

A reference table for phases
}
\description{
create a list of phases in any project

create a list of phases in any project

create a list of phases in any project
}
