% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-project-teams.R
\name{set_project_team}
\alias{set_project_team}
\title{Assigns staff roles on a project}
\usage{
set_project_team(project_teams)
}
\arguments{
\item{project_teams}{Data frame that lists project, role, staff, assigned_capacity}
}
\value{
A reference table project teams and roles
}
\description{
Assigns staff roles on a project
}
