% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-project_teams.R
\name{set_project_team}
\alias{set_project_team}
\title{Assigns staff roles on a project}
\usage{
set_project_team(projects, roles, assigned_staff = NA,
  assigned_capacity = NA)
}
\arguments{
\item{projects}{Data frame that lists project, role, staff, assigned_capacity}

\item{roles}{vvv}

\item{assigned_capacity}{vvv}

\item{staff}{nnn}
}
\value{
A reference table project teams and roles
}
\description{
Assigns staff roles on a project
}
\examples{
staff <- c('Shelby', 'Luis', 'Taishawn', 'Samantha', 'Taylor')
capacity <- c(40,60,100,100,100)
resources <- set_resources(staff, capacity)
projects <- LETTERS[1:3]
probability <- c(50, 100, 100)
start <- as.Date(c("2019-07-25", "2019-05-17", "2019-09-27")) 
end <- as.Date(c("2019-09-03", "2019-06-16", "2019-10-27"))
projects <- set_projects(projects, probability, start, end)
roles <- c("lead", "researcher", "editor", "design")
roles <- set_roles(roles)
staff <- sample(resources$staff, 
                              size = nrow(project_teams), 
                              replace = TRUE)
assigned_capacity <- sample(c(25,25,75,100), 
                                   size = nrow(project_teams), 
                                   replace = TRUE)
project_teams <- set_project_team(project, roles, staff, assigned_capacity)
}
