% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-project_teams.R
\name{set_project_team}
\alias{set_project_team}
\title{Assigns staff roles on a project}
\usage{
set_project_team(projects, roles, assigned_staff = NA,
  assigned_capacity = NA)
}
\arguments{
\item{projects}{Data frame that lists project, role, staff, assigned_capacity}

\item{roles}{vvv}

\item{assigned_capacity}{vvv}

\item{staff}{nnn}
}
\value{
A reference table project teams and roles
}
\description{
Assigns staff roles on a project
}
