% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-set_projects.R
\name{set_projects}
\alias{set_projects}
\title{create a list of employees that are to be assigned to projects}
\usage{
set_projects(projects, probability = 1, start, end)
}
\arguments{
\item{projects}{Names of projects}

\item{probability}{Probability that the project will go ahead}

\item{start}{Expected start date of project}

\item{end}{Expected end date of the project}
}
\value{
A reference table for projects
}
\description{
create a list of employees that are to be assigned to projects
}
\keyword{internal}
