% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-set_holidays.R
\name{set_public_holidays}
\alias{set_public_holidays}
\title{create a list of phases in any project}
\usage{
set_public_holidays(date, name)
}
\arguments{
\item{file}{file with public holiday information}
}
\value{
A reference table for phases
}
\description{
create a list of phases in any project
}
\keyword{internal}
