% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-time_estimates.R
\name{set_time_estimates}
\alias{set_time_estimates}
\title{creates time estimates for each phases of each project}
\usage{
set_time_estimates(projects, phases, days_from_end)
}
\arguments{
\item{projects}{List of roles in any project team in order of responsibility}

\item{phases}{List of roles in any project team in order of responsibility}

\item{days_from_end}{Time estimates of how long each phase will take in relation to project end, negative = phase will occur before project end, positive = phase will occur after project end}
}
\value{
A reference table for time estimates of each phase of each project
}
\description{
creates time estimates for each phases of each project
}
\examples{
projects <- LETTERS[1:3]
probability <- c(50, 100, 100)
start <- as.Date(c("2019-07-25", "2019-05-17", "2019-09-27")) 
end <- as.Date(c("2019-09-03", "2019-06-16", "2019-10-27"))
projects <- set_projects(projects, probability, start, end)
phases <- c("research", "drafting", "editing", "design", "print", "events")
phases <- set_phases(phases)
time_estimates <- rbind(c(-40,-10,-10,-10,-10,10), c(-10,-10,-10,0,0,0), 
                        c(0,0,0,-10,-10,10))
time_estimates <- set_time_estimates(projects, phases, time_estimates)
}
