% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-set_classes.R
\docType{class}
\name{time_estimates-class}
\alias{time_estimates-class}
\alias{workplanr_time_estimates}
\title{Base Class for workplan time_estimates}
\description{
Base Class for workplan time_estimates
}
\section{Slots}{

\describe{
\item{\code{project_name}}{List of roles in any project team in order of responsibility}

\item{\code{project_phase_name}}{List of roles in any project team in order of responsibility}

\item{\code{time_estimates}}{Time estimates of how long each phase will take in relation 
to project end, negative = phase will occur before project end, positive = phase will occur after project end}
}}

\seealso{
Other classes: \code{\link{out_of_office-class}},
  \code{\link{project_assignments-class}},
  \code{\link{project_phases-class}},
  \code{\link{project_roles-class}},
  \code{\link{projects-class}},
  \code{\link{public_holidays-class}},
  \code{\link{roles_responsibilities-class}},
  \code{\link{schedule-class}}, \code{\link{staff-class}},
  \code{\link{staff_schedule-class}},
  \code{\link{team_schedule-class}},
  \code{\link{workplan-class}}
}
\concept{classes}
