% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-workplanr.R
\name{workplanr-package}
\alias{workplanr-package}
\alias{iep.charts-package}
\title{IEP charting functions for R}
\description{
\code{workplanr} package
}
\details{
The \code{workplanr} package is a SQLite dabase wrapper to plan concurrent projects with phases and deadlines.
It in intended for project manaers who need to track dealdlines and stff loads across multiple projects.

It is aimed to be a lightweight simple to use projec management tool to avoid the need for small teams to invest 
in large expensive project management tool infrastructure.

To build a workplan, the user needs to set the following project inputs
\itemize{
\item \strong{staff} A data frame of staff_name and staff capacity
\item Producing finalised charts for layout
\item Producing maps in IEP colours
}

These functions are intended to produce pdfs that are importable into illustrator
making the layout for comms much easier.
}
\section{Installation}{


#' # Then you need devtools installed

\code{install.packages('devtools')}

# Now you can install from github 

\code{devtools::install_github('hammond/workplar')}

\code{library(workplanr)}
}

\author{
David Hammond \email{anotherdavidhammond@gmail.com}
}
\keyword{package}
